+++ test/test_commands/test_scenario12.py
@@ -12,6 +12,7 @@ def test_yaml_it_src_to_out_12_powershell():
         uncompiled = Path("scenario12_ps1/folder")
         output_root = Path("scenario12_ps1/out")
         shutil.rmtree(str(Path(__file__).parent / "scenario12_ps1/out"), ignore_errors=True)
+        shutil.rmtree(str(Path(__file__).parent / "scenario12_ps1/folder/.bash2gitlab"), ignore_errors=True)
 
         run_compile_all(uncompiled, output_root)