+++ test/test_commands/test_scenario13.py
@@ -12,6 +12,7 @@ def test_yaml_it_src_to_out_13_complex_data_structure():
         uncompiled = Path("scenario13/src")
         output_root = Path("scenario13/out")
         shutil.rmtree(str(Path(__file__).parent / "scenario13/out"), ignore_errors=True)
+        shutil.rmtree(str(Path(__file__).parent / "scenario13/src/.bash2gitlab"), ignore_errors=True)
 
         run_compile_all(uncompiled, output_root)