+++ test/test_commands/test_scenario7.py
@@ -12,6 +12,7 @@ def test_yaml_it_src_to_out_7():
         uncompiled = Path("scenario7/in")
         output_root = Path("scenario7/out")
         shutil.rmtree(str(Path(__file__).parent / "scenario7/out"), ignore_errors=True)
+        shutil.rmtree(str(Path(__file__).parent / "scenario7/in/.bash2gitlab"), ignore_errors=True)
 
         run_compile_all(uncompiled, output_root)